# analyze: borrowck performance improvements